third_party/WebKit/LayoutTests/fast/dom/global-event-handlers.html

<!DOCTYPE html>
<title>GlobalEventHandlers test</title>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script>
// attribute list from WHATWG HTML Living Standard r8212
var attributes = [
    "onabort",
    "onblur",
    "onerror",
    "onfocus",
    "oncancel",
    "oncanplay",
    "oncanplaythrough",
    "onchange",
    "onclick",
    "onclose",
    "oncontextmenu",
    "oncuechange",
    "ondblclick",
    "ondrag",
    "ondragend",
    "ondragenter",
    "ondragexit",
    "ondragleave",
    "ondragover",
    "ondragstart",
    "ondrop",
    "ondurationchange",
    "onemptied",
    "onended",
    "oninput",
    "oninvalid",
    "onkeydown",
    "onkeypress",
    "onkeyup",
    "onload",
    "onloadeddata",
    "onloadedmetadata",
    "onloadstart",
    "onmousedown",
    "onmouseenter",
    "onmouseleave",
    "onmousemove",
    "onmouseout",
    "onmouseover",
    "onmouseup",
    "onmousewheel",
    "onpause",
    "onplay",
    "onplaying",
    "onprogress",
    "onratechange",
    "onreset",
    "onscroll",
    "onseeked",
    "onseeking",
    "onselect",
    "onshow",
    "onsort",
    "onstalled",
    "onsubmit",
    "onsuspend",
    "ontimeupdate",
    "onvolumechange",
    "onwaiting"
];
function getObject(interface) {
    switch(interface) {
        case "Element":
            var e = document.createElementNS("http://example.com/", "example");
            assert_true(e instanceof Element);
            assert_false(e instanceof HTMLElement);
            assert_false(e instanceof SVGElement);
            return e;
        case "HTMLElement":
            var e = document.createElement("html");
            assert_true(e instanceof HTMLElement);
            return e;
        case "SVGElement":
            var e = document.createElementNS("http://www.w3.org/2000/svg", "svg");
            assert_true(e instanceof SVGElement);
            return e;
        case "Document":
            assert_true(document instanceof Document);
            return document;
        case "Window":
            assert_true(window instanceof Window);
            return window;
    }
    assert_unreached();
}
function testSet(interface, attribute) {
    test(function() {
        var object = getObject(interface);
        function nop() {}
        assert_equals(object[attribute], null, "Initially null");
        object[attribute] = nop;
        assert_equals(object[attribute], nop, "Return same function");
        document[attribute] = "";
        assert_equals(document[attribute], null, "Return null after setting string");
    }, "Set " + interface + "." + attribute);
}
function testReflect(interface, attribute) {
   test(function() {
        var element = getObject(interface);
        assert_equals(element.getAttribute(attribute), null, "Initially null");
        element.setAttribute(attribute, "return");
        assert_equals(element.getAttribute(attribute), "return", "Return same string");
        assert_equals(typeof element[attribute], "function", "Convert to function");
    }, "Reflect " + interface + "." + attribute);
}
// Object.propertyIsEnumerable cannot be used because it doesn't
// work with properties inherited through the prototype chain.
var enumerable = {
    "Element": {},
    "HTMLElement": {},
    "SVGElement": {},
    "Document": {},
    "Window": {}
};
Object.keys(enumerable).forEach(function(interface) {
    for (var attribute in getObject(interface)) {
        enumerable[interface][attribute] = true;
    }
});
function testEnumerate(interface, attribute) {
    test(function() {
        assert_true(enumerable[interface][attribute]);
    }, "Enumerate " + interface + "." + attribute);
}
attributes.forEach(function(attribute) {
    test(function() {
        assert_false(attribute in getObject("Element"));
    }, "No Element." + attribute);
    testSet("HTMLElement", attribute);
    testEnumerate("HTMLElement", attribute);
    testReflect("HTMLElement", attribute);
    testSet("Document", attribute);
    testEnumerate("Document", attribute);
    testSet("Window", attribute);
    testEnumerate("Window", attribute);
});
</script>
<div id="log"></div>
